% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_criterion.R
\name{avg_dev}
\alias{avg_dev}
\title{Helper to compute the mean absolute deviation of a validation set}
\usage{
avg_dev(valid_set, alpha)
}
\arguments{
\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
the average absolute deviation.
}
\description{
Helper to compute the mean absolute deviation of a validation set
}
\keyword{internal}
