% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_criterion.R
\name{avg_width}
\alias{avg_width}
\title{Helper to compute the mean conformal interval length of a validation set}
\usage{
avg_width(valid_set, alpha)
}
\arguments{
\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
the average conformal interval length.
}
\description{
Helper to compute the mean conformal interval length of a validation set
}
\keyword{internal}
