% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_criterion.R
\name{conf_homo}
\alias{conf_homo}
\title{Helper to compute confident homogeneity}
\usage{
conf_homo(width, deviation, lambda)
}
\arguments{
\item{width}{(\code{number})\cr total conformal interval length of a split.}

\item{deviation}{(\code{number})\cr total absolute deviation of a split.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
the value of the confident homogeneity.
}
\description{
Helper to compute confident homogeneity
}
\keyword{internal}
