% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_homogeneity.R
\name{conf_quantile}
\alias{conf_quantile}
\title{Helper to compute the conformal alpha-quantile for a vector of residuals}
\usage{
conf_quantile(residuals, alpha)
}
\arguments{
\item{residuals}{(\code{numeric})\cr residuals of the predictions on the validation set.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
the conformal alpha-quantile.
}
\description{
Helper to compute the conformal alpha-quantile for a vector of residuals
}
\keyword{internal}
