% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_node.R
\name{crit_node}
\alias{crit_node}
\title{Helper to compute the confident homogeneity in a node.}
\usage{
crit_node(valid_set, alpha, lambda)
}
\arguments{
\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
the confident homogeneity in the node.
}
\description{
Helper to compute the confident homogeneity in a node.
}
\keyword{internal}
