% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_splitting.R
\name{eval_split_cand}
\alias{eval_split_cand}
\title{Helper to assess if a split is sensible}
\usage{
eval_split_cand(ids_left, valid_set, alpha)
}
\arguments{
\item{ids_left}{(\code{integer})\cr ids of the left child node.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
Logical value indicating if the child nodes contain enough observations.
}
\description{
This checks if the child nodes contain enough observations to result in sensible quantile values.
This is to prevent that the alpha-quantile becomes \eqn{\infty} if \eqn{(1 - \alpha)(n + 1) > n}.
}
\keyword{internal}
