% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{get_homogeneity}
\alias{get_homogeneity}
\title{Helper to get confident homogeneity for all inner and outer nodes in a tree}
\usage{
get_homogeneity(tree, valid_set, alpha, lambda)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
List of the confident homogeneity in the tree's nodes.
}
\description{
Helper to get confident homogeneity for all inner and outer nodes in a tree
}
\keyword{internal}
