% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_homogeneity.R
\name{get_pred_mean}
\alias{get_pred_mean}
\title{Helper to compute the mean prediction of the validation set}
\usage{
get_pred_mean(valid_set)
}
\arguments{
\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}
}
\value{
the mean of the predicted target variable.
}
\description{
Helper to compute the mean prediction of the validation set
}
\keyword{internal}
