% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_growing.R
\name{get_split}
\alias{get_split}
\title{Helper to get the next split of a tree}
\usage{
get_split(candidates, x_data)
}
\arguments{
\item{candidates}{(\code{list})\cr all sensible splits in the tree's terminal nodes.
See \code{\link[=get_candidates]{get_candidates()}} for details.}

\item{x_data}{(\code{data.frame})\cr feature data matrix.}
}
\value{
List specifying the next split
}
\description{
Helper to get the next split of a tree
}
\keyword{internal}
