% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_set.R
\name{get_valid_set}
\alias{get_valid_set}
\title{Helper to retrieve validation set and predictions}
\usage{
get_valid_set(data, target, learner, cv_folds, treatment = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr.}

\item{target}{(\code{string})\cr name of the target variable.}

\item{learner}{(\code{model_spec} object)\cr the learner for training the prediction model.
See \code{\link[parsnip:model_spec]{parsnip::model_spec()}} for details.}

\item{cv_folds}{(\code{count})\cr number of CV+ folds.}

\item{treatment}{(\code{string})\cr name of the treatment variable, if applicable.}
}
\value{
Data frame with predictions and indices.
}
\description{
Helper to retrieve validation set and predictions
}
\keyword{internal}
