% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_set.R
\name{get_validation_set}
\alias{get_validation_set}
\title{Helper to Retrieve Validation Set and Predictions}
\usage{
get_validation_set(data, target, learner, cv_folds)
}
\arguments{
\item{data}{(\code{data.frame})\cr.}

\item{target}{(\code{string})\cr name of the target variable.}

\item{learner}{(\code{model_spec} object)\cr the learner for training the prediction model.
See \code{\link[parsnip:model_spec]{parsnip::model_spec()}} for details.}

\item{cv_folds}{(\code{count})\cr number of CV+ folds.}
}
\value{
Data frame with predictions and indices.
}
\description{
Helper to Retrieve Validation Set and Predictions
}
\keyword{internal}
