% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot.conftree}
\alias{plot.conftree}
\title{Plot a party tree}
\usage{
\method{plot}{conftree}(x, ...)
}
\arguments{
\item{x}{(\code{conftree})\cr tree containing detected subgroups.}

\item{...}{additional arguments.}
}
\value{
The plot.
}
\description{
Plot a party tree
}
