% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{print.conftree}
\alias{print.conftree}
\title{Print a conftree}
\usage{
\method{print}{conftree}(x, ...)
}
\arguments{
\item{x}{(\code{conftree})\cr tree containing detected subgroups.}

\item{...}{additional arguments.}
}
\description{
Print a conftree
}
\examples{
library(tidymodels)
library(ranger)
data(bikes)
set.seed(1234)
randforest <- rand_forest(trees = 200, min_n = 5) \%>\%
  set_mode("regression") \%>\%
  set_engine("ranger")
groups <- r2p(
  data = bikes,
  target = "count",
  learner = randforest,
  cv_folds = 2,
  alpha = 0.05,
  gamma = 0.2,
  lambda = 0.5,
  max_groups = 3
)
print(groups)
}
