% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_splitting.R
\name{process_covariate}
\alias{process_covariate}
\title{Helper to find all sensible splits in a covariate}
\usage{
process_covariate(
  var_name,
  x_data,
  node_id,
  valid_set,
  crit_node,
  alpha,
  lambda
)
}
\arguments{
\item{var_name}{(\code{string})\cr name of the feature to be split.}

\item{x_data}{(\code{data.frame})\cr feature data matrix.}

\item{node_id}{(\code{count})\cr parent node identifier.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{crit_node}{(\code{number})\cr confident criterion of the parent node.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
List of sensible splits in the covariate.
}
\description{
Helper to find all sensible splits in a covariate
}
\keyword{internal}
