% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_splitting.R
\name{process_node}
\alias{process_node}
\title{Helper to find all sensible splits in a parent node}
\usage{
process_node(x_data, node_id, valid_set, alpha, gamma, lambda)
}
\arguments{
\item{x_data}{(\code{data.frame})\cr feature data matrix.}

\item{node_id}{(\code{count})\cr parent node identifier.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{gamma}{(\code{proportion})\cr regularization parameter ensuring that reduction
in the impurity of the confident homogeneity is sufficiently large.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
List of all sensible splits in the parent node.
}
\description{
Helper to find all sensible splits in a parent node
}
\keyword{internal}
