% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_splitting.R
\name{process_split_config}
\alias{process_split_config}
\title{Helper to compute the gain in confident criterion of a split}
\usage{
process_split_config(ids_left, ids_right, valid_set, crit_node, alpha, lambda)
}
\arguments{
\item{ids_left}{(\code{integer})\cr ids of the left child node.}

\item{ids_right}{(\code{integer})\cr ids of the right child node.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{crit_node}{(\code{number})\cr confident criterion of the parent node.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
The value of the confident criterion of a split.
}
\description{
Helper to compute the gain in confident criterion of a split
}
\keyword{internal}
