% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_splitting.R
\name{process_split_config_numeric}
\alias{process_split_config_numeric}
\title{Helper to compute the confident criterion of a numeric split}
\usage{
process_split_config_numeric(
  node_id,
  var_name,
  split_cand,
  x_data,
  valid_set,
  alpha,
  lambda
)
}
\arguments{
\item{node_id}{(\code{count})\cr parent node identifier.}

\item{var_name}{(\code{string})\cr name of the feature to be split.}

\item{split_cand}{(\code{number})\cr value of the split point.}

\item{x_data}{(\code{data.frame})\cr feature data matrix.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
List with elements \code{node_id}, \code{feature}, \code{feature_type}, \code{split_cand}, \code{ids_left_child}, \code{ids_right_child} and \code{conf_crit}.
}
\description{
Helper to compute the confident criterion of a numeric split
}
\keyword{internal}
