% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{r2p-package}
\alias{r2p}
\alias{r2p-package}
\title{\code{r2p} Package}
\description{
\code{r2p} detects subgroups via recursive partitioning based on confident conformal homogeneity.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
