% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{round_plot}
\alias{round_plot}
\title{Helper to round numbers according to their magnitude}
\usage{
round_plot(x, sig = 2)
}
\arguments{
\item{x}{(\code{numeric})\cr number to round.}

\item{sig}{(\code{number})\cr non-zero decimal digits to consider for rounding.}
}
\value{
Rounded number.
}
\description{
Helper to round numbers according to their magnitude
}
\keyword{internal}
