% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_criterion.R
\name{total_width}
\alias{total_width}
\title{Helper to compute the total conformal interval length of a candidate split}
\usage{
total_width(valid_set, ids_left, ids_right, alpha)
}
\arguments{
\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{ids_left}{(\code{integer})\cr ids of the left child node.}

\item{ids_right}{(\code{integer})\cr ids of the right child node.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
the total conformal interval length.
}
\description{
Helper to compute the total conformal interval length of a candidate split
}
\keyword{internal}
