% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{tree_dev}
\alias{tree_dev}
\title{Helper to get absolute deviation for all inner and outer nodes in a tree}
\usage{
tree_dev(tree, valid_set, alpha)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}
}
\value{
List of the absolute deviation in the tree's nodes.
}
\description{
Helper to get absolute deviation for all inner and outer nodes in a tree
}
\keyword{internal}
