% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{tree_predmean}
\alias{tree_predmean}
\title{Helper to get the mean prediction for all inner and outer nodes in a tree}
\usage{
tree_predmean(tree, valid_set)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}
}
\value{
List of the mean prediction in the tree's nodes.
}
\description{
Helper to get the mean prediction for all inner and outer nodes in a tree
}
\keyword{internal}
