% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{tree_var_nodes}
\alias{tree_var_nodes}
\title{Helper to get the variances within a tree's nodes}
\usage{
tree_var_nodes(tree, valid_set, terminal = TRUE)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{terminal}{(\code{flag})\cr if \code{TRUE} compute only for terminal nodes.}
}
\value{
List of the variance within the tree's nodes.
}
\description{
Helper to get the variances within a tree's nodes
}
\keyword{internal}
