% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{tree_var_within}
\alias{tree_var_within}
\title{Helper to get the variance within a tree's terminal nodes}
\usage{
tree_var_within(tree, valid_set)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}
}
\value{
Variance within the tree's terminal nodes.
}
\description{
Helper to get the variance within a tree's terminal nodes
}
\keyword{internal}
