% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tree.R
\name{tree_vars}
\alias{tree_vars}
\title{Helper to get the variances across and within a tree's terminal nodes}
\usage{
tree_vars(tree, target)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{target}{(\code{string})\cr name of the target variable. The target must be a numeric variable.}
}
\value{
Variance across and within the tree's terminal nodes.
}
\description{
Helper to get the variances across and within a tree's terminal nodes
}
\keyword{internal}
