% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conftree-package.R, R/package.R
\docType{package}
\name{conftree-package}
\alias{conftree}
\alias{conftree-package}
\title{conftree: Subgroups with Trees using Conformal Prediction}
\description{
Detect subgroups in machine learning models with numeric outcomes (regression). This leverages CV+/Jackknife+ to simultaneously optimize inter-group heterogeneity and intra-group homogeneity. In this way, it is an extension of the r2p algorithm, see Lee et al. (2020) \href{https://arxiv.org/abs/2006.07917v2}{arXiv:2006.07917v2}. This works with any of the learning algorithms for regression tasks available in 'tidymodels'.

\code{conftree} detects subgroups via recursive partitioning based on confident conformal homogeneity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://holgstr.github.io/conftree/}
}


Useful links:
\itemize{
  \item \url{https://holgstr.github.io/conftree/}
}

}
\author{
\strong{Maintainer}: Holger Löwe \email{hbj.loewe@gmail.com}

}
\keyword{internal}
