% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{conftree-package}
\alias{conftree}
\alias{conftree-package}
\title{\code{conftree} Package}
\description{
\code{conftree} detects subgroups via recursive partitioning based on confident conformal homogeneity.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
