% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_growing.R
\name{get_candidates}
\alias{get_candidates}
\title{Helper to get candidates for the next split of a tree}
\usage{
get_candidates(tree, x_data, valid_set, alpha, gamma, lambda)
}
\arguments{
\item{tree}{(\code{party})\cr tree object.
See \code{\link[partykit:party]{partykit::party()}} for details.}

\item{x_data}{(\code{data.frame})\cr feature data matrix.}

\item{valid_set}{(\code{data.frame})\cr validation set.
See \code{\link[=get_valid_set]{get_valid_set()}} for details.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{gamma}{(\code{proportion})\cr regularization parameter ensuring that reduction
in the impurity of the confident homogeneity is sufficiently large.}

\item{lambda}{(\code{proportion})\cr balance between width and deviation.}
}
\value{
List of all sensible splits in the tree's terminal nodes.
}
\description{
Helper to get candidates for the next split of a tree
}
\keyword{internal}
