% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2p.R
\name{r2p}
\alias{r2p}
\title{Finding Subgroups With Regression Trees Using Conformal Prediction}
\usage{
r2p(
  data,
  target,
  learner,
  cv_folds = 2,
  alpha = 0.05,
  gamma = 0.01,
  lambda = 0.5
)
}
\arguments{
\item{data}{(\code{data.frame})\cr data set for model training and uncertainty estimation.}

\item{target}{(\code{string})\cr name of the target variable.}

\item{learner}{(\code{model_spec})\cr the learner for training the prediction model.
See \code{\link[parsnip:model_spec]{parsnip::model_spec()}} for details.}

\item{cv_folds}{(\code{count})\cr number of CV+ folds.}

\item{alpha}{(\code{proportion})\cr miscoverage rate.}

\item{gamma}{(\code{proportion})\cr regularization parameter ensuring that reduction
in the impurity of the confident homogeneity is sufficiently large.}

\item{lambda}{(\code{proportion})\cr balance parameter, quantifying the impact of the average interval length relative
to the average absolute deviation (i.e. interval size vs. within-group homogeneity)}
}
\value{
The tree.
}
\description{
Finding Subgroups With Regression Trees Using Conformal Prediction
}
\examples{
library(tidymodels)
library(ranger)
data(bikes)
set.seed(1234)
randforest <- rand_forest(trees = 200, min_n = 5) \%>\%
 set_mode("regression") \%>\%
 set_engine("ranger")
r2p(data = bikes,
     target = "count",
     learner = randforest,
     cv_folds = 2,
     alpha = 0.05,
     gamma = 0.2,
     lambda = 0.5)
}
